refactor unittests + more unittests
