+++ app/ilmoweb/views.py
@@ -635,9 +635,12 @@ def teachers_message(request):
         View for updating message in instuctions page
     """
     if request.method == "GET":
-        if not request.user.is_staff:
-            return redirect("/open_labs")
-        return redirect("/system")
+        if not request.user.is_superuser:
+            if request.user.is_staff:
+                return redirect(created_labs)
+            if not request.user.is_staff:
+                return redirect(open_labs)
+        return redirect(system)
 
     if request.method == "POST":
         new_message = request.POST.get("message")
